% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ges_mpi.R,
%   R/seurat_sample_expression_2.R
\name{extract_ges_mpi}
\alias{extract_ges_mpi}
\title{Title}
\usage{
extract_ges_mpi(
  aggregate_out_dir = ".",
  min_n_reads_per_cluster = 20000,
  n_replicates = 5,
  outdir = "NETZEN_analysis",
  organism = "human",
  clusters_file = "./analysis/clustering/graphclust/clusters.csv",
  extract_sample_cluster = FALSE
)

extract_ges_mpi(
  aggregate_out_dir = ".",
  min_n_reads_per_cluster = 20000,
  n_replicates = 5,
  outdir = "NETZEN_analysis",
  organism = "human",
  clusters_file = "./analysis/clustering/graphclust/clusters.csv",
  extract_sample_cluster = FALSE
)
}
\arguments{
\item{extract_sample_cluster}{}

\item{extract_cell_group}{}
}
\description{
Title

Title
}
