% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_sample_expression_2.R
\name{extract_ges_mpi_old}
\alias{extract_ges_mpi_old}
\title{Title}
\usage{
extract_ges_mpi_old(
  aggregate_out_dir = ".",
  min_n_reads_per_cluster = 20000,
  n_replicates = 5,
  outdir = "NETZEN_analysis",
  organism = "human",
  clusters_file = "./analysis/clustering/graphclust/clusters.csv",
  extract_sample_cluster = FALSE
)
}
\arguments{
\item{extract_sample_cluster}{}
}
\description{
Title
}
