% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ges_parallel.R
\name{extract_ges_parallel}
\alias{extract_ges_parallel}
\title{Title}
\usage{
extract_ges_parallel(
  tasks,
  min_n_reads_per_cluster = 20000,
  outdir = "NETZEN_analysis",
  organism = "human",
  method = "mpi"
)
}
\arguments{
\item{method}{method of parallel processing, choose from :mpi, multicore, singlecore.}
}
\description{
Title
}
