% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ges.R
\name{extract_task_ges}
\alias{extract_task_ges}
\title{Title}
\usage{
extract_task_ges(task, ges, n_replicates = 5)
}
\arguments{
\item{task}{the task to calculate expression, a list of barcodes, n_replicates, barcode_group_name}

\item{min_n_cells_per_cluster}{}
}
\value{
a list ( result: list of (barcodes_ges_sum dataframe, barcodes_ges_sum_replicates dataframe, is_null_result), task_id)
}
\description{
Title
}
