% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_aggregation_sample_gene_expressions}
\alias{get_aggregation_sample_gene_expressions}
\title{Title
Extract aggregated gene expression of samples in scRNA seq, convert from single cell RNAseq to bulk RNAseq.}
\usage{
get_aggregation_sample_gene_expressions(
  aggregate_out_dir,
  aggregation_samples_file,
  outdir,
  organism,
  method = "mpi",
  close_slaves = TRUE,
  n_replicates = 5,
  min_n_reads_per_cell_group = 20000
)
}
\arguments{
\item{aggregate_out_dir}{aggregate output directory from 10X run, parent directory of "filtered_feature_bc_matrix"}

\item{aggregation_samples_file}{aggregation sample file for 10X pipeline, containing columns: library_id, molecule_h5 and other metadata column}

\item{outdir}{output directory}

\item{organism}{either mouse or human . Use to translate gene ENSM id into normal gene names}

\item{method}{method to run parallel processing. Choose from "mpi", "multicore", "singlecore"}

\item{close_slaves}{default = TRUE. if method is mpi, then close slaves node}

\item{n_replicates}{number of replicates to sample from}

\item{min_n_reads_per_cell_group}{minimal number of reads per cell group (sample)}
}
\value{
write in output directory the pseudo bulk RNA seq with following format: count, tpm, rpkm, cpm.
}
\description{
Title
Extract aggregated gene expression of samples in scRNA seq, convert from single cell RNAseq to bulk RNAseq.
}
