% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_expression_from_tasks.R
\name{get_cell_group_gene_expression_from_tasks}
\alias{get_cell_group_gene_expression_from_tasks}
\title{Title}
\usage{
get_cell_group_gene_expression_from_tasks(
  tasks,
  aggregate_out_dir,
  outdir,
  prefix,
  organism,
  method = "mpi",
  close_slaves = TRUE,
  n_replicates = 5,
  min_n_reads_per_cell_group = 20000
)
}
\description{
Title
}
