% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_cell_group_gene_expressions_from_annotation}
\alias{get_cell_group_gene_expressions_from_annotation}
\title{Title
Extract aggregated gene expression of cell group in scRNA seq, convert from single cell RNAseq to bulk RNAseq. Each cell group is a cell group in annotated file.}
\usage{
get_cell_group_gene_expressions_from_annotation(
  aggregate_out_dir,
  annotation_file,
  outdir,
  organism,
  method = "mpi",
  close_slaves = TRUE,
  n_replicates = 5,
  min_n_reads_per_cell_group = 20000,
  prefix = "cell_group"
)
}
\arguments{
\item{aggregate_out_dir}{aggregate output directory from 10X run, parent directory of "filtered_feature_bc_matrix"}

\item{annotation_file}{aggregation sample file for 10X pipeline, containing columns: library_id, molecule_h5 and other metadata column}

\item{outdir}{output directory}

\item{organism}{either mouse or human . Use to translate gene ENSM id into normal gene names}

\item{method}{method to run parallel processing. Choose from "mpi", "multicore", "singlecore"}

\item{close_slaves}{default = TRUE. if method is mpi, then close slaves node}

\item{n_replicates}{number of replicates to sample from}

\item{min_n_reads_per_cell_group}{minimal number of reads per cell group (sample_cluster)}

\item{prefix}{prefix to add to the output subdirectories and files. For example: "immunocell_annotation"}
}
\description{
Title
Extract aggregated gene expression of cell group in scRNA seq, convert from single cell RNAseq to bulk RNAseq. Each cell group is a cell group in annotated file.
}
