% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_expression_data.R
\name{get_formated_ges_from_counts}
\alias{get_formated_ges_from_counts}
\title{Title
from count data frame of cell group gene expression, filter cell groups that have number of read higher than threshold, and the total number of reads for each cell group}
\usage{
get_formated_ges_from_counts(
  cell_group_ges,
  gene_names,
  min_n_reads_per_cell_group = 20000
)
}
\arguments{
\item{cell_group_ges}{a count dataframe where first column is ENS id, other columns are cell group names, rows are genes, values are counts}

\item{gene_names}{the corresponding normal gene names}

\item{min_n_reads_per_cell_group}{Minimal number of reads per cell group threshold to filter}
}
\value{
a list of:
cell_group_ges:original count data frame
cell_group_ges_cpm : cpm gene expression dataframe
cell_group_ges_hugo :count data frame with normal gene names instead of ENS id
cell_group_ges_hugo_cpm : cpm gene expression data frame with normal gene names instead of ENS id
filtered_cell_group_ges :  filtered count data frame, where cell groups have total number of reads more than min_n_reads_per_cell_group
filtered_cell_group_ges_hugo :  filtered count data frame with normal gene names instead of ENS id, where cell groups have total number of reads more than min_n_reads_per_cell_group
total_n_reads_per_cell_group : total number of reads per cell group data frame with two columns:  total_n_reads_per_cell_group,  cell_group.
total_n_reads_per_cell_group_more_than_threshold :  total number of reads per cell group that pass threshold data frame with two columns:  total_n_reads_per_cell_group,  cell_group.
}
\description{
Title
from count data frame of cell group gene expression, filter cell groups that have number of read higher than threshold, and the total number of reads for each cell group
}
