% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{get_normalized_data}
\alias{get_normalized_data}
\title{Title}
\usage{
get_normalized_data(counts = NULL, organism = "human")
}
\arguments{
\item{organism}{Name of organism such as "human" or "mouse". Used for translation from ENSEMBL ID to gene name.}
}
\description{
Title
}
