% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_generator.R
\name{get_task_from_aggregation_samples}
\alias{get_task_from_aggregation_samples}
\title{Title}
\usage{
get_task_from_aggregation_samples(
  aggregation_samples_file = "aggregation.csv",
  barcode.names
)
}
\description{
Title
}
