% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ges_parallel.R
\name{load_ges_slave}
\alias{load_ges_slave}
\title{Title}
\usage{
load_ges_slave(aggregate_out_dir, n_replicates = 5)
}
\arguments{
\item{aggregate_out_dir}{Aggregation outdir, parent folder of folder filtered_feature_bc_matrix}

\item{n_replicates}{number of replicates to sample from each sing cell group}
}
\description{
Title
}
