% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_expression_data.R
\name{normalize_and_write_result}
\alias{normalize_and_write_result}
\title{Title
Save the result of get_formated_ges_from_counts function to disk , also get normalized gene expression by calling the normalized_and_export_counts_data function}
\usage{
normalize_and_write_result(
  formated_ges_list,
  outdir,
  prefix = "sample",
  organism = "human"
)
}
\arguments{
\item{formated_ges_list}{the result of get_formated_ges_from_counts}

\item{outdir}{output directory}

\item{prefix}{prefix to add to the file path, such as sample, sample_replicates, sample_cluster, sample_cluster_duplicates, etc.}

\item{organism}{either mouse or human . Use to translate gene ENSM id into normal gene names}
}
\description{
Title
Save the result of get_formated_ges_from_counts function to disk , also get normalized gene expression by calling the normalized_and_export_counts_data function
}
