% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalized_and_export_counts_data}
\alias{normalized_and_export_counts_data}
\title{Title}
\usage{
normalized_and_export_counts_data(
  counts = NULL,
  outdir = ".",
  organism = "human"
)
}
\arguments{
\item{organism}{}
}
\description{
Title
}
