% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNetZen-package.R
\docType{package}
\name{scNetZen-package}
\alias{scNetZen}
\alias{scNetZen-package}
\title{scNetZen: Seurat Analysis Pipeline}
\description{
Extract gene expression from single cell
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sonbangle/scNetZen}
  \item Report bugs at \url{https://github.com/sonbangle/scNetZen/issues}
}

}
\author{
\strong{Maintainer}: Son Le \email{nuibang@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
