% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpi_control.R
\name{slave_control}
\alias{slave_control}
\title{Title}
\usage{
slave_control(FUN, dot.arg = NULL)
}
\arguments{
\item{dot.arg:}{additional argument list for function, equivalent to ...}
}
\description{
Title
}
