% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_func.R
\name{sonicApply}
\alias{sonicApply}
\title{Title
A parallel version of lapply using either mpi on clusters, multicore or simple lapply
Advatange comparing with Rmpi mpi.applyBL: More flexible, can do some preprocessing in the slave node first, for example reading gene expression in each
slave node, not passing the big share data from master to slaves.}
\usage{
sonicApply(x, fun, method, close_slaves = TRUE, ...)
}
\arguments{
\item{x}{a list of elements to process}

\item{fun}{function to apply}

\item{method}{method to do parallel processing, either mpi, multicore or single core}

\item{close_slaves}{close slave after doing parallel apply. default is TRUE, set to FALSE when you want to reuse slave.}
}
\description{
Title
A parallel version of lapply using either mpi on clusters, multicore or simple lapply
Advatange comparing with Rmpi mpi.applyBL: More flexible, can do some preprocessing in the slave node first, for example reading gene expression in each
slave node, not passing the big share data from master to slaves.
}
