% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpi_control.R
\name{TaskApply}
\alias{TaskApply}
\title{Title
Parallelly apply the function to list, similar to lapply but with mpi version
Initialization and closing mpi should be done manually so that each slave can be prepared(loading the dataset) before processing the data}
\usage{
TaskApply(tasks, fun, ...)
}
\arguments{
\item{tasks}{the list of data, each element is a data to apply function}

\item{fun}{The name of function on the slave. Should be exist in slave node either by creating on the node or transfer from the master manually}
}
\value{
the list of concatentated output, result of applying function into each input element
}
\description{
Title
Parallelly apply the function to list, similar to lapply but with mpi version
Initialization and closing mpi should be done manually so that each slave can be prepared(loading the dataset) before processing the data
}
