% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{write_cell_group_gene_expression_from_counts}
\alias{write_cell_group_gene_expression_from_counts}
\title{Title
From cell group counts data frame, write different cell group gene_expression normalized data, and filtered for the cell groups with total numbers of reads above threshold}
\usage{
write_cell_group_gene_expression_from_counts(
  cell_group_counts,
  gene_names,
  outdir,
  prefix,
  organism,
  min_n_reads_per_cell_group = 20000
)
}
\arguments{
\item{cell_group_counts}{count dataframe where first column is ENS id, other columns are cell group names, rows are genes, values are counts}

\item{gene_names}{the corresponding normal gene names}

\item{outdir}{output directory}

\item{prefix}{prefix to add to the file path, such as sample, sample_replicates, sample_cluster, sample_cluster_duplicates, etc.}

\item{organism}{either mouse or human . Use to translate gene ENSM id into normal gene names}

\item{min_n_reads_per_cell_group}{Minimal number of reads per cell group threshold to filter}
}
\description{
Title
From cell group counts data frame, write different cell group gene_expression normalized data, and filtered for the cell groups with total numbers of reads above threshold
}
