% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ges.R
\name{extract_ges}
\alias{extract_ges}
\title{Title
Extract gene expression from group of cells, such as a cluster of cells, the whole sample or any group of cells}
\usage{
extract_ges(barcodes, ges, barcode_group_name, n_replicates = 5)
}
\arguments{
\item{barcodes}{the barcodes of single cells in the group of interest.}

\item{ges}{gene expression profile}

\item{barcode_group_name}{barcode group name}

\item{n_replicates}{number of replicates when do sampling of cells in the group}

\item{task_id}{tak_id}

\item{min_n_cells}{minimal number of cells in the group to calculate}
}
\description{
Title
Extract gene expression from group of cells, such as a cluster of cells, the whole sample or any group of cells
}
