% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ges_parallel.R
\name{extract_ges_parallel}
\alias{extract_ges_parallel}
\title{Title}
\usage{
extract_ges_parallel(
  tasks,
  method = "mpi",
  close_slaves = TRUE,
  n_replicates = 5
)
}
\arguments{
\item{method}{method of parallel processing, choose from :mpi, multicore, singlecore.}
}
\description{
Title
}
